Updated Data Retrieval from FAOstats
We have recently updated our data retrieval process from FAOstats. The changes include new and improved links to ensure seamless and efficient data extraction.

New Forecasts for Corn

In addition, we are excited to announce that we have added new forecasts for Corn. These forecasts incorporate the latest data for the year 2022, ensuring that our users have access to the most current and accurate information available.